refactor: consolidate EntityType tests into ServiceTest class and remove old EntityTypeTests class
